refactor: matchInfo 컴포넌트를 matchCard로 변경
