+++ src/components/common/MatchCard/pieces/Status.tsx
@@ -1,4 +1,4 @@
-import { useMatchInfoContext } from '@/hooks/useMatchInfoContext';
+import { useMatchCardContext } from '@/hooks/useMatchCardContext';
 import { $ } from '@/utils/core';
 
 type StatusProps = {
@@ -6,7 +6,7 @@ type StatusProps = {
 };
 
 export default function Status({ className }: StatusProps) {
-  const { gameQuarter } = useMatchInfoContext();
+  const { gameQuarter } = useMatchCardContext();
 
   return (
     <div className={$('z-10 px-5 text-white', className)}>{gameQuarter}</div>
