+++ src/hooks/useMatchCardContext.ts
@@ -1,11 +1,11 @@
 import { useContext } from 'react';
 
-import { MatchContext } from '@/components/common/MatchInfo/pieces/Wrapper';
+import { MatchContext } from '@/components/common/MatchCard/pieces/Wrapper';
 import { MatchDetailType } from '@/types/match';
 
-type MatchInfoContextType = () => MatchDetailType;
+type MatchCardContextType = () => MatchDetailType;
 
-export const useMatchInfoContext: MatchInfoContextType = () => {
+export const useMatchCardContext: MatchCardContextType = () => {
   const matchContext = useContext(MatchContext);
 
   if (!matchContext) throw new Error('Context가 비었습니다.');
